Add files via upload
This project is an analysis of a dataset from SL(Stockholms Länstrafik). This is the first part of the project, where I extract the variables that will be used. I am also cleaning up the data by using imputation on the missing values. The method is called Simple random imputation, witch means that that the missing values are replaced by values that are subtracted from a probability distribution function based on the remaining data.